requirements.txt
PyExecJS==1.5.1
PySimpleGUI==4.60.5